Typecast to double to avoid rounding off

+++ traffic/trafficGenerator.cpp
@@ -48,7 +48,7 @@ int DcTrafficGenerator::loadTrafficDistribution (std::string packetsizeDistFile,
     pkt_size_t base_size = 80; // in bytes
     pkt_size_t size_on_wire = base_size + 24;
     int pps = ((this->torLinkSpeed * 1000000000)/ (size_on_wire * 8));
-    sim_time_t min_delay_ns = (1/pps * 1000000000);
+    sim_time_t min_delay_ns = ((double)1/(double)pps * 1000000000);
 
 
     myRandomFromCDF.loadCDFs("traffic-dist/fb_webserver_packetsizedist_cdf.csv", "traffic-dist/fb_webserver_flowinterarrival_ns_cdf.csv", min_delay_ns);
